azure-connector/pyproject.toml
[project]
name = "azure-connector"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "azure-iot-device>=2.14.0",
    "dotenv>=0.9.9",
]
